Introducing plaster, a nice little python script
It plasters whatever matching wallpapers it can on screens that match
the aspect ratio, and is of lower resolution than the screen. It's
currently quite limited, and you have to run it manually after sway has
been active for a bit. I suspect this has to do with the outputs not
being initialized upon sway start.